LapVet/requirements.txt
astroid==2.3.3
autopep8==1.4.4
branca==0.3.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.2.9
django-ckeditor==5.8.0
django-js-asset==1.2.2
django-multiselectfield==0.1.11
folium==0.10.1
html5lib==1.0.1
idna==2.8
isort==4.3.21
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.0
pandas==0.24.2
Pillow==6.2.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.4
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tinycss2==1.0.2
typed-ast==1.4.0
urllib3==1.25.7
WeasyPrint==51
webencodings==0.5.1
wrapt==1.11.2